+++ common/src/main/java/fr/rakambda/fallingtree/common/tree/breaking/InstantaneousTreeBreakingHandler.java
@@ -3,7 +3,6 @@ package fr.rakambda.fallingtree.common.tree.breaking;
 import fr.rakambda.fallingtree.common.FallingTreeCommon;
 import fr.rakambda.fallingtree.common.tree.Tree;
 import fr.rakambda.fallingtree.common.tree.TreePart;
-import fr.rakambda.fallingtree.common.wrapper.ILevel;
 import fr.rakambda.fallingtree.common.wrapper.IPlayer;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -16,6 +15,7 @@ public class InstantaneousTreeBreakingHandler implements ITreeBreakingHandler{
 	private static InstantaneousTreeBreakingHandler INSTANCE;
 	
 	private final FallingTreeCommon<?> mod;
+	private final LeafForceBreaker leafForceBreaker;
 	
 	@Override
 	public boolean breakTree(@NotNull IPlayer player, @NotNull Tree tree) throws BreakTreeTooBigException{
@@ -61,35 +61,15 @@ public class InstantaneousTreeBreakingHandler implements ITreeBreakingHandler{
 		}
 		
 		if(brokenCount >= toolHandler.getMaxBreakCount()){
-			forceBreakDecayLeaves(player, tree, level);
+			leafForceBreaker.forceBreakDecayLeaves(player, tree, level);
 		}
 		return true;
 	}
 	
-	private void forceBreakDecayLeaves(@NotNull IPlayer player, @NotNull Tree tree, @NotNull ILevel level){
-		var radius = mod.getConfiguration().getTrees().getLeavesBreakingForceRadius();
-		if(radius > 0){
-			tree.getTopMostLog().ifPresent(topLog -> {
-				var start = topLog.offset(-radius, -radius, -radius);
-				var end = topLog.offset(radius, radius, radius);
-				topLog.betweenClosedStream(start, end).forEach(checkPos -> {
-					var checkState = level.getBlockState(checkPos);
-					var checkBlock = checkState.getBlock();
-					if(mod.isLeafBlock(checkBlock)){
-						if(!player.isCreative() || mod.getConfiguration().isLootInCreative()){
-							checkState.dropResources(level, checkPos);
-						}
-						level.removeBlock(checkPos, false);
-					}
-				});
-			});
-		}
-	}
-	
 	@NotNull
-	public static InstantaneousTreeBreakingHandler getInstance(@NotNull FallingTreeCommon<?> common){
+	public static InstantaneousTreeBreakingHandler getInstance(@NotNull FallingTreeCommon<?> mod){
 		if(isNull(INSTANCE)){
-			INSTANCE = new InstantaneousTreeBreakingHandler(common);
+			INSTANCE = new InstantaneousTreeBreakingHandler(mod, new LeafForceBreaker(mod));
 		}
 		return INSTANCE;
 	}
